backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
firebase-admin
google-cloud-firestore
python-multipart
PyPDF2
sentence-transformers
numpy
pydantic
